kmeans seems to still be generating empty clusters. To mitigate this, I've added a new metric that replaces empty clusters, with vectors that are furthest from their centroids.

+++ src/main/java/edu/uc/rphash/Centroid.java
@@ -6,7 +6,7 @@ import edu.uc.rphash.Readers.RPVector;
 
 public class Centroid {
 	public float[] centroid;
-	public long count;
+	public float count;
 	public int dimensions;
 	public ConcurrentSkipListSet<Long> ids;
 	public long id;
@@ -177,7 +177,7 @@ public class Centroid {
 //	this.wcss = M2;//tmpsum / (float) count;
 	}
 
-	public Long getCount() {
+	public float getCount() {
 		return count;
 	}
